requirement.txt
requests==2.31.0
allure-pytest==2.13.5
pymysql==1.1.0
loguru==0.7.2
pandas==2.2.2
openpyxl==3.1.2
selenium==4.21.0